feat: choose submodule (#132)
Fixes #49 

# Summary

Implements the Choose Submodule flow for constructs with submodules. This is currently limited to submodules that are a single level deep.

![Screen Shot 2021-06-23 at 1 41 06 PM](https://user-images.githubusercontent.com/8749859/123153832-c349ad80-d42b-11eb-8d0b-a72681029c86.png)
![Screen Shot 2021-06-23 at 1 41 30 PM](https://user-images.githubusercontent.com/8749859/123153837-c47ada80-d42b-11eb-9a7f-fc9d6214bfea.png)
![Screen Shot 2021-06-23 at 1 42 41 PM](https://user-images.githubusercontent.com/8749859/123153838-c47ada80-d42b-11eb-885f-eb7ae7629125.png)

I prototyped a version with a modal rather than a popover which always rendered in the center of the viewport. Unfortunately the pic is cropped around the component, but the key difference is that it always renders mid-viewport, and has a dark overlay, creating a stronger contrast from the backgrounded content. Sharing below as I can't really decide which is the better UX. I believe the modal is a more a11y friendly UX, though it strays from the wireframes

![Screen Shot 2021-06-22 at 5 31 26 PM](https://user-images.githubusercontent.com/8749859/123162174-e8dbb480-d435-11eb-9119-8d4859211ecd.png)