feat: catalog search component (#146)
Prerequisite to #139 

This PR implements a responsive search component with state controlled externally and a language filter. Once we have the capability to filter on CDK type, a small update will be needed to add that functionality. This component is tested in storybook and unit tests, and will be integrated into the app in a follow-up to reduce PR size.

This work also co-locates all component stories with their respective components

![Screen Shot 2021-06-28 at 5 22 02 PM](https://user-images.githubusercontent.com/8749859/123711400-0b0a7380-d836-11eb-9f8d-433a22a04e9d.png)
![Screen Shot 2021-06-28 at 5 22 07 PM](https://user-images.githubusercontent.com/8749859/123711401-0b0a7380-d836-11eb-93a7-3afa2eb6f2b9.png)