chore: setting up for mobile package nav (#184)
This work sets up the package page to be ready for mobile nav functionality. The main change is that the language bar has been moved to the package details card, and the side nav is hidden on mobile. In a follow up, the side nav will be rendered in the mobile nav drawer.

Additionally, this PR moves the `CatalogSearchInputs` component into a separate file per [this comment](https://github.com/cdklabs/construct-hub-webapp/pull/180#discussion_r668119441) 

![Screen Shot 2021-07-12 at 3 54 27 PM](https://user-images.githubusercontent.com/8749859/125355814-175cf900-e32b-11eb-9064-7c1f2cf78ccc.png)
![Screen Shot 2021-07-12 at 3 54 41 PM](https://user-images.githubusercontent.com/8749859/125355818-17f58f80-e32b-11eb-874b-8e95e4acef5e.png)