chore: fetch markdown from backend  (#183)
Following https://github.com/cdklabs/construct-hub/pull/162, we should now be fetching the markdown from the backend instead of generating it in the client. 

In addition, switch usages of `reflect.Assembly` (i.e `jsii-reflect`) to `spec.Assembly` (i.e `@jsii/spec`) since it is not required anymore and `spec.Assembly` is an easier type to work with.