feat: better configure ReactMarkdown
Enable remark-gfm so that all GitHub-Flag Markdown is supported,
including tables (Fixes cdklabs/construct-hub#197).

Properly sanitize all generated markdown using `rehype-sanitize`,
ensuring we do not render anything that GitHub would not.

Transform image and link URLs that are repository-relative, so that such
links are valid. These resolve to GitHub distribution points in the same
way that npmjs.com does, so this change is neutral for packages that are
not configured with a GitHub repository (Fixes cdklabs/construct-hub#198
cdklabs/construct-hub#199).