fix: missing doc pages result in an unclear error  (#260)
Currently, when users browse to a documentation page in a language that is either:

1. Unsupported by the package
2. Documentation isn't ready for yet
3. Documentation failed to create for

This is what we show:

![Screen Shot 2021-07-25 at 11 36 11 AM](https://user-images.githubusercontent.com/1428812/126893012-8e23d701-be4d-4c8d-97de-0d3d56605f92.png)

> Example: https://constructs.dev/packages/@softchef/cdk-vue/v/1.0.31?lang=python

This happens because we currently don't detect that we requested a missing document, and we simply render the `index.html` page that CloudFront serves in such scenarios. 

This PR adds this missing docs page detection by asserting that the content type of the response is the one we expect. 
It works because our doc pages have `text/markdown`, while our `index.html` has `text/html`. 

This is just a tactical stop-gap until we implement proper 404 responses for such cases on the backend. 

Following this PR, these errors will displayed as:

![Screen Shot 2021-07-25 at 11 26 42 AM](https://user-images.githubusercontent.com/1428812/126893094-60985395-b3f1-4f7b-bf1c-a00095eb99a3.png)

The error still needs some design, but at least it's clear and coherent. 

Fixes https://github.com/cdklabs/construct-hub-webapp/issues/255
Fixes https://github.com/cdklabs/construct-hub/issues/218