feat: pageview analytics (#251)
Fixes https://github.com/cdklabs/construct-hub-webapp/issues/9

This draft implements pageview analytics for our application. When a user navigates to a page, an analytics call will be made which capture the page name, url, and event name. This PR is going into draft as it will need reviews from some of the EU folks who are away until monday.

As for the implementation:
I've built two primary components for the analytics API:
1. AnalyticsProvider - provides an analytics API which can be expanded as-needed in the future
2. Page - a generic page component which views can render to capture pageview analytics. The API is very simple at this time, but it is designed to be expanded for other types of page-behaviors if needed