fix: long api references do not load (#287)
An example would be: https://constructs.dev/packages/cdk-iam-floyd/v/0.218.0

This PR puts a hard limit of 1MB on the size of the docs page, anything above that will cause only the README to be displayed, along with a notice that the API reference couldn't be rendered. 

Note that the README displayed is the transliterated README, so at least we have that added value for those packages. 

This is just a stop-gap until we are able to resolve the issue, but at least we won't crash and stall browsers. 

![Screen Shot 2021-07-29 at 11 29 55 AM](https://user-images.githubusercontent.com/1428812/127467014-2367b6b5-3c6f-444f-b707-d2eb301d3f9e.png)
![Screen Shot 2021-07-29 at 12 15 53 PM](https://user-images.githubusercontent.com/1428812/127467051-c883c949-34e8-4af4-9962-4929312d80c9.png)