fix: dependency links are broken (#293)
Dependency links didn't remove the `^` from the version number. Now we sanitize the version string and removing any leading `^` or `~`. 

In addition, now that we have multiple major versions (this was actually true before as well) - the package latest redirection was broken for those package. Now it will redirect to the latest major version.