fix: page crash if preferred language is not available
The installation instructions fail to render if the default language
configured by the user is not available for the current package (i.e:
this particulkar target is not present for the jsii assembly thereof).

Instead of crashing, this renders a message telling the user this
language is not available, and that they should enagge the author if
they are interested in it.