fix: packaging
Add packaging step back into release workflow so publish to npm isn't
broken.

This was initially removed because we are pushing the src code in the
npm tarball in addition to the build output to allow build time
configuration if needed, but the projen release workflow relies on the
output of the packaging step and can't publish succesfully without it.