fix: home page not rendering when package has no keywords (#449)
I found a surprising bug when testing the website against my updated instance of the Construct Hub, where the home page simply went blank because of an error "w.keywords is undefined" coming from CatalogCard.tsx. This is caused because packages are now ingested on the backend even if they might not have CDK keywords (https://github.com/cdklabs/construct-hub/pull/330), so the "keywords" field may be null. This fixes the error by changing the type of keywords to possibly be undefined.

I also added a projen task for updating snapshots, as well as a package to the catalog.json fixture so that we have a reference example of packages without any keywords.