chore: fetch package data from network instead of cache by default (#553)
Should fix #512

Updating caching policies implemented by our service worker which is being used to support PWA functionality such as using the website offline.

Currently, if the user has visited the website previously but they have old data cached, the service worker will request data "from both the cache and the network in parallel" per the `StaleWhileRevalidate` policy. This will update the cache with the newest data, but this data will only be served on subsequent responses, which explains existing behavior where the user only sees new data after a refresh.

This PR changes updates the service worker so that dynamic data sources such as the package catalog will first attempt to return the response from the network, and only fall back to the cache if no response is received. In doing so the app should still work gracefully in offline scenarios.