fix: replace service worker with no-op worker (#623)
In #571 we updated the default JavaScript code to try and unregister service workers. However, the same service-worker.js file is still being served at the root of the website, which may cause browsers that still have the old service worker installed to think there has been no changes to the service worker, so it may continue serving old cached content. A more proper fix as described by [this Stackoverflow post](https://stackoverflow.com/questions/33986976/how-can-i-remove-a-buggy-service-worker-or-implement-a-kill-switch/38980776) is to replace the service worker with a no-op worker that does not try serving any cached content.