feat: categories section in home page
Add a "categories" section in the home page with a set of buttons that lead to keyword searches.

If the backend config includes a `categories` section, it will be used to define the set of categories and keywords. If it is absent, a default set will be included.

The new design also switches the colors of the Featured and Tabs section in the home page.