fix: routes do not redirect properly (#690)
Fixes #688 

The original bug is that we were comparing the versions as strings, so it thought version "10" < "7" and hence redirected us to v7.

The correct behavior we actually want is to link to the most recently published version of the package, to match the behavior of linking to the latest version of the package on search results in #643. (Long term, we want this to redirect to whichever version of the package has the "latest" NPM dist tag).