fix: multiple of same package appears in search results (#643)
Fixes #585

Instead of showing the major version of a package (or major version of the CDK framework) to the package cards, we filter out packages appearing multiple times with different major versions from search results, always preferring to show the highest major version of the package that satisfies the search filters.

For example, if you search for `@aws-cdk/cx-api`, by default only one card is shown, that links to `@aws-cdk/cx-api@2` that supports CDK v2. But if you apply a filter for AWS CDK v1, then `@aws-cdk/cx-api@1` is shown.