fix: api ref nav behaves differently from readme
When clicking on a nav item from readme, it will scroll the view so that this item is at the top (which implies that the package header is not displayed). When clicking on an API reference item, it will scroll to 0 (which implies that package header is displayed). This results in an inconsistent exprience.

To solve this, we added an `id` to the API doc header and scroll it to view if the URL is an `/api` URL.

Additionally, we also made this header a link, so that it will behave like README headers (which are links to themselves).