fix(shortbread): cookie consent not showing on Safari, Chrome
The ShortBread initializatio code is registered in a `window.load` event
handler, however it appears that event has already fired in WebKit-like
browsers when the event handler is added.

Just initialize ShortBread in-band instead, so it always shows.

Additionally - adjusted to dev-mode CSP to match the production site's
with the required adjustments (dev-mode is served over HTTP, and so most
of the sources are also HTTP instead of HTTPS like on production).

Also, fixed a React warning pertaining to a `forwardRef` use.