chore: unpin chakra-ui version (#798)
A few weeks ago an automatic dependency upgrade introduced a bugged version of chakra-ui which would not allow a fixed color mode (it always inherited the user's system color mode). This bug has since been patched, so we are safe to keep up with the latest versions of chakra-ui now.