feat: support preload script (#879)
This pull request updates the webapp index.html to point to the new `preload.js` file rather than hard-coding the scripts in our public library. This will allow Construct Hub operators to define and use their own preload scripts, which can serve as an entry point for monitoring or analytics.

### Testing (localhost)

1. `git fetch && git checkout feat/preload-script`
2. `yarn && yarn dev`
3. Open the network tab, verify that `preload.js` loads and that analytic calls are made by filtering `b/ss`