chore: update TOC to be fixed width (#884)
Updates the table of contents so that it always has the same width on desktop viewports, as it can seem overly large or overly squished on certain screen resolutions.

Also updates the table of contents to display ellipses when titles get too long -- the CSS for ellipses was already there, but the text was still wrapping since the `white-space` property wasn't set to `nowrap`.

<img width="454" alt="Screen Shot 2022-01-27 at 4 16 08 PM" src="https://user-images.githubusercontent.com/5008987/151464908-4bae34e4-9676-4b11-a76f-68cdb14d4d68.png">