feat: class navigation sidebar (#895)
Closes #319

Adds a class navigation sidebar for quickly navigating methods, properties, etc.

To avoid too much squishing because of the extra horizontal room it takes up, I changed the responsive layout for package pages so that the switching point between tablet and desktop layouts is at chakra's "lg" instead of "md".

Currently this sidebar is created based on the JSON doc schema that is generated in conjunction with markdown docs for each package. Since these are different sources of information, there's some logic that had to be duplicated from how we are generating the markdown and generating links / IDs. Once we generate the API reference straight from the API.json and remove the markdown from the picture, it's possible we could simplify this down a lot.

Screenshot:

<img width="1287" alt="Screen Shot 2022-02-01 at 3 33 55 PM" src="https://user-images.githubusercontent.com/5008987/152069248-93039309-0581-4d36-89a4-b3b17cde15a5.png">