feat: move readme subitems to right nav (#915)
Fixes #914 

This pull request moves README subitems from the primary doc navigation to the secondary nav. The first commit on this pull request re-organizes logical code from presentational code, while the second commit actually does the work to implement the feature.

While testing this feature, I've also observed that libraries which are missing readmes (common for submodules) have an unclear UX and we should try to add some messaging to clarify that scenario i.e "This {library or submodule} does not have a readme. Use the menu on the left to browse the API Reference"

![Screenshot 2022-02-15 at 11-36-28 projen 0 52 31 - Construct Hub](https://user-images.githubusercontent.com/8749859/154136494-fae037c5-a464-4a84-8f99-5f76a7b5798f.png)