fix: badge color broken in safari (#1002)
On Safari, our Construct Hub badge SVG has a coloring issue:

<img width="379" alt="Screen Shot 2022-06-14 at 3 45 57 PM" src="https://user-images.githubusercontent.com/5008987/173701642-883257a4-b48c-45b7-b40f-21912d1784dd.png">

After some googling I initially thought this was caused by a bug in WebKit (https://bugs.webkit.org/show_bug.cgi?id=199134) -- but after diving a little deeper, it turns out this is caused by an erroneous line of code in the SVG. In Chrome/Firefox it seems that this erroneous line doesn't make a difference, but in Safari it causes coloring issues.

For consistency I've also updated the non-dynamically-colored SVGs with the same fix.